# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/ip/{ip}": {
            "get": {
                "description": "IP查询",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP查询"
                ],
                "summary": "IP查询",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IP address",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/model.IPInfoResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.IPInfoResponse": {
            "type": "object",
            "properties": {
                "addr": {
                    "type": "string"
                },
                "asn": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "latitude": {
                    "type": "string"
                },
                "longitude": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "registered_country": {
                    "type": "string"
                },
                "subdivisions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}